refactor: improve API and code quality
