+++ src/registry/types/ChainRegistryChainInfo.ts
@@ -1,0 +1,227 @@
+/* eslint-disable */
+/**
+ * This file was automatically generated by json-schema-to-typescript.
+ * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
+ * and run json-schema-to-typescript to regenerate this file.
+ */
+
+/**
+ * Cosmos Chain.json is a metadata file that contains information about a cosmos sdk based chain.
+ */
+export interface ChainRegistryChainInfo {
+  $schema?: string;
+  chain_name: string;
+  chain_id: string;
+  pre_fork_chain_name?: string;
+  pretty_name?: string;
+  website?: string;
+  update_link?: string;
+  status?: "live" | "upcoming" | "killed";
+  network_type?: "mainnet" | "testnet" | "devnet";
+  /**
+   * The default prefix for the human-readable part of addresses that identifies the coin type. Must be registered with SLIP-0173. E.g., 'cosmos'
+   */
+  bech32_prefix: string;
+  /**
+   * Used to override the bech32_prefix for specific uses.
+   */
+  bech32_config?: {
+    /**
+     * e.g., 'cosmos'
+     */
+    bech32PrefixAccAddr?: string;
+    /**
+     * e.g., 'cosmospub'
+     */
+    bech32PrefixAccPub?: string;
+    /**
+     * e.g., 'cosmosvaloper'
+     */
+    bech32PrefixValAddr?: string;
+    /**
+     * e.g., 'cosmosvaloperpub'
+     */
+    bech32PrefixValPub?: string;
+    /**
+     * e.g., 'cosmosvalcons'
+     */
+    bech32PrefixConsAddr?: string;
+    /**
+     * e.g., 'cosmosvalconspub'
+     */
+    bech32PrefixConsPub?: string;
+  };
+  daemon_name?: string;
+  node_home?: string;
+  key_algos?: ("secp256k1" | "ethsecp256k1" | "ed25519" | "sr25519")[];
+  slip44?: number;
+  alternative_slip44s?: number[];
+  fees?: {
+    fee_tokens: FeeToken[];
+  };
+  staking?: {
+    staking_tokens: StakingToken[];
+    lock_duration?: {
+      /**
+       * The number of blocks for which the staked tokens are locked.
+       */
+      blocks?: number;
+      /**
+       * The approximate time for which the staked tokens are locked.
+       */
+      time?: string;
+    };
+  };
+  codebase?: {
+    git_repo?: string;
+    recommended_version?: string;
+    compatible_versions?: string[];
+    binaries?: {
+      "linux/amd64"?: string;
+      "linux/arm64"?: string;
+      "darwin/amd64"?: string;
+      "darwin/arm64"?: string;
+      "windows/amd64"?: string;
+      "windows/arm64"?: string;
+    };
+    cosmos_sdk_version?: string;
+    consensus?: {
+      type: "tendermint" | "cometbft";
+      version?: string;
+    };
+    cosmwasm_version?: string;
+    cosmwasm_enabled?: boolean;
+    /**
+     * Relative path to the cosmwasm directory. ex. $HOME/.juno/data/wasm
+     */
+    cosmwasm_path?: string;
+    ibc_go_version?: string;
+    /**
+     * List of IBC apps (usually corresponding to a ICS standard) which have been enabled on the network.
+     */
+    ics_enabled?: ("ics20-1" | "ics27-1" | "mauth")[];
+    genesis?: {
+      name?: string;
+      genesis_url: string;
+      ics_ccv_url?: string;
+    };
+    versions?: {
+      /**
+       * Official Upgrade Name
+       */
+      name: string;
+      /**
+       * Git Upgrade Tag
+       */
+      tag?: string;
+      /**
+       * Block Height
+       */
+      height?: number;
+      /**
+       * Proposal that will officially signal community acceptance of the upgrade.
+       */
+      proposal?: number;
+      /**
+       * [Optional] Name of the following version
+       */
+      next_version_name?: string;
+      recommended_version?: string;
+      compatible_versions?: string[];
+      cosmos_sdk_version?: string;
+      consensus?: {
+        type: "tendermint" | "cometbft";
+        version?: string;
+      };
+      cosmwasm_version?: string;
+      cosmwasm_enabled?: boolean;
+      /**
+       * Relative path to the cosmwasm directory. ex. $HOME/.juno/data/wasm
+       */
+      cosmwasm_path?: string;
+      ibc_go_version?: string;
+      /**
+       * List of IBC apps (usually corresponding to a ICS standard) which have been enabled on the network.
+       */
+      ics_enabled?: ("ics20-1" | "ics27-1" | "mauth")[];
+      binaries?: {
+        "linux/amd64"?: string;
+        "linux/arm64"?: string;
+        "darwin/amd64"?: string;
+        "darwin/arm64"?: string;
+        "windows/amd64"?: string;
+        "windows/arm64"?: string;
+      };
+    }[];
+  };
+  images?: {
+    image_sync?: Pointer;
+    png?: string;
+    svg?: string;
+    theme?: {
+      primary_color_hex?: string;
+    };
+  }[];
+  logo_URIs?: {
+    png?: string;
+    svg?: string;
+  };
+  peers?: {
+    seeds?: Peer[];
+    persistent_peers?: Peer[];
+  };
+  apis?: {
+    rpc?: Endpoint[];
+    rest?: Endpoint[];
+    grpc?: Endpoint[];
+    wss?: Endpoint[];
+    "grpc-web"?: Endpoint[];
+    "evm-http-jsonrpc"?: Endpoint[];
+  };
+  explorers?: Explorer[];
+  keywords?: string[];
+  extra_codecs?: ("ethermint" | "injective")[];
+}
+export interface FeeToken {
+  denom: string;
+  fixed_min_gas_price?: number;
+  low_gas_price?: number;
+  average_gas_price?: number;
+  high_gas_price?: number;
+  gas_costs?: {
+    cosmos_send?: number;
+    ibc_transfer?: number;
+  };
+}
+export interface StakingToken {
+  denom: string;
+}
+/**
+ * The (primary) key used to identify an object within the Chain Registry.
+ */
+export interface Pointer {
+  /**
+   * The chain name or platform from which the object resides. E.g., 'cosmoshub', 'ethereum', 'forex', or 'nasdaq'
+   */
+  chain_name: string;
+  /**
+   * The base denom of the asset from which the object originates. E.g., when describing ATOM from Cosmos Hub, specify 'uatom', NOT 'atom' nor 'ATOM'; base units are unique per platform.
+   */
+  base_denom?: string;
+}
+export interface Peer {
+  id: string;
+  address: string;
+  provider?: string;
+}
+export interface Endpoint {
+  address: string;
+  provider?: string;
+  archive?: boolean;
+}
+export interface Explorer {
+  kind?: string;
+  url?: string;
+  tx_page?: string;
+  account_page?: string;
+}
