+++ src/registry/types/KeplrChainInfo.ts
@@ -1,0 +1,3 @@
+import type { ChainInfo as KeplrChainInfo } from "@keplr-wallet/types";
+
+export type { KeplrChainInfo };
